requirements.txt
python-telegram-bot==21.7
python-dotenv==1.0.0
sqlalchemy>=2.0.35
flask==3.0.0
flask-wtf==1.2.1
alembic==1.13.0
